refactor: :construction: Restructuring the Demands page
Restructuring the Demands page
