+++ src/Components/Molecules/Table/index.module.scss
