+++ src/Components/Templates/DemandasTemplate/index.module.scss
@@ -3,18 +3,6 @@
   grid-template-columns: repeat(2, 1fr);
   grid-auto-flow: row dense;
   grid-auto-rows: min-content;
-  gap: 1rem; 
+  gap: 1rem;
   padding: 1rem;
 }
-  
-.dashSectionLarge {
-  grid-row: span 2; 
-  margin-bottom: 1rem;
-  }
-  
-.dashSectionSmall {
-  grid-column: span 1;
-  grid-row: span 0,5;
-  margin-bottom: 1rem;
-}
-
